backend/requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fsspec==2024.12.0
greenlet==3.1.1
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
six==1.17.0
SQLAlchemy==2.0.37
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
